requirements.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple

# CUDA 12.1 官方 wheels
torch==2.6.0+cu124          # ➜ 对应 torchvision 0.18 / torchaudio 2.3
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124

# 其余依赖
timm>=0.9.14
albumentations>=1.4.0
opencv-python-headless>=4.10
torchmetrics>=1.3.2
pyyaml>=6.0.1

numpy~=1.26.4
pillow~=10.3.0
pandas~=2.2.3
tqdm~=4.67.1
stringzilla~=3.12.5
typing_extensions~=4.13.2
pip~=25.0.1
filelock~=3.18.0
pytz~=2025.2
safetensors~=0.5.3
requests~=2.32.3
colorama~=0.4.6
setuptools~=78.1.0
scipy~=1.15.2
mpmath~=1.3.0
sympy~=1.13.1
packaging~=25.0
urllib3~=2.4.0
fsspec~=2025.3.2
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
certifi~=2025.1.31
simsimd~=6.2.1
albucore~=0.0.23
six~=1.16.0
networkx~=3.4.2
pydantic~=2.11.3
pydantic_core~=2.33.1
idna~=3.10
Jinja2~=3.1.6
kornia~=0.8.0